added changes

+++ src/main/java/com/example/posassist/dto/request/OrderItemDTO.java
@@ -12,5 +12,5 @@ public class OrderItemDTO {
 
     private String instructions;
 
-    private Integer quantity = 1;
+    private Double quantity = 1.0;
 }
